requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.0.5
openai==0.28.0
python-dotenv==1.0.0